Split up AAChannelProcessor for clarity
