+++ src/main/java/org/phoebus/channelfinder/processors/aa/ArchiveAction.java
@@ -1,0 +1,18 @@
+package org.phoebus.channelfinder.processors.aa;
+
+enum ArchiveAction {
+    ARCHIVE("/archivePV"),
+    PAUSE("/pauseArchivingPV"),
+    RESUME("/resumeArchivingPV"),
+    NONE("");
+
+    private final String endpoint;
+
+    ArchiveAction(final String endpoint) {
+        this.endpoint = endpoint;
+    }
+
+    public String getEndpoint() {
+        return this.endpoint;
+    }
+}
