+++ src/test/java/org/phoebus/channelfinder/processors/AAChannelProcessorNoDefaultIT.java
@@ -10,6 +10,7 @@ import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.phoebus.channelfinder.entity.Channel;
 import org.phoebus.channelfinder.entity.Property;
+import org.phoebus.channelfinder.processors.aa.AAChannelProcessor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.test.context.TestPropertySource;
@@ -20,7 +21,6 @@ import java.util.stream.Stream;
 
 import static org.phoebus.channelfinder.processors.AAChannelProcessorIT.activeProperty;
 import static org.phoebus.channelfinder.processors.AAChannelProcessorIT.archiveProperty;
-import static org.phoebus.channelfinder.processors.AAChannelProcessorIT.inactiveProperty;
 import static org.phoebus.channelfinder.processors.AAChannelProcessorIT.paramableAAChannelProcessorTest;
 
 @WebMvcTest(AAChannelProcessor.class)
